fix(general-petition): cleanup in project (#10145)
