--- libs/application/templates/general-petition/src/graphql/mutations.ts
@@ -1,16 +1,0 @@
-import gql from 'graphql-tag'
-
-export const EndorseList = gql`
-  mutation endorsementSystemEndorseList($input: CreateEndorsementInput!) {
-    endorsementSystemEndorseList(input: $input) {
-      id
-      endorser
-      endorsementListId
-      meta {
-        fullName
-      }
-      created
-      modified
-    }
-  }
-`
